requirements.txt
# فریمورک جنگو
django==4.2

# خواندن متغیرهای محیطی از فایل .env
python-decouple

# درایور PostgreSQL به همراه باینری‌های مورد نیاز
psycopg[binary]

# فریمورک REST API جنگو
djangorestframework

# JWT برای احراز هویت توکن‌محور
djangorestframework-simplejwt

# فیلتر کردن داده‌ها در APIها
django-filter

# تولید مستندات Swagger برای API
drf-yasg

# کار با تصاویر (مثلا آپلود عکس)
Pillow

requests

instaloader